fix: survive OOM and read errors on four failure paths
- ChapterHtmlSlimParser: heading and <li> marker no longer dereference a
  text block that failed to allocate.
- PageImage::deserialize returns nullptr when its ImageBlock is missing
  instead of building an image around a null block.
- KOSync and nearby position sync check the Epub allocation.
- KOReaderDocumentId reads into an int: a -1 read error became SIZE_MAX
  as size_t and was hashed.

+++ lib/KOReaderSync/KOReaderDocumentId.cpp
@@ -76,11 +76,12 @@ std::string KOReaderDocumentId::calculate(const std::string& filePath) {
 
     // Read up to CHUNK_SIZE bytes
     const size_t bytesToRead = std::min(CHUNK_SIZE, fileSize - offset);
-    const size_t bytesRead = file.read(buffer, bytesToRead);
+    // int, not size_t: a read error is -1, which as size_t hashed SIZE_MAX bytes.
+    const int bytesRead = file.read(buffer, bytesToRead);
 
     if (bytesRead > 0) {
-      md5.add(buffer, bytesRead);
-      totalBytesRead += bytesRead;
+      md5.add(buffer, static_cast<size_t>(bytesRead));
+      totalBytesRead += static_cast<size_t>(bytesRead);
     }
   }
